Split implementation into smaller files
