+++ component/mimir/rules/rules.go
@@ -1,23 +1,17 @@
 package rules
 
 import (
-	"bytes"
 	"context"
 	"fmt"
 	"time"
 
-	"github.com/ghodss/yaml"
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 	"github.com/grafana/agent/component"
 	mimirClient "github.com/grafana/agent/pkg/mimir/client"
 	"github.com/grafana/dskit/crypto/tls"
-	"github.com/grafana/dskit/multierror"
-	"github.com/pkg/errors"
 	promv1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
 	promListers "github.com/prometheus-operator/prometheus-operator/pkg/client/listers/monitoring/v1"
-	"github.com/prometheus/prometheus/model/rulefmt"
-	yamlv3 "gopkg.in/yaml.v3"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
@@ -139,12 +133,6 @@ func (c *Component) Update(newConfig component.Arguments) error {
 	return <-errChan
 }
 
-func setDefaultArguments(args *Arguments) {
-	if args.SyncInterval == 0 {
-		args.SyncInterval = 30 * time.Second
-	}
-}
-
 func (c *Component) init() error {
 	level.Info(c.log).Log("msg", "initializing with new configuration")
 
@@ -211,33 +199,6 @@ func (c *Component) init() error {
 	return nil
 }
 
-func (c *Component) reconcileState(ctx context.Context) error {
-	ctx, cancel := context.WithTimeout(ctx, 5*time.Second)
-	defer cancel()
-
-	crdState, err := c.ruleLister.List(c.ruleSelector)
-	if err != nil {
-		return fmt.Errorf("failed to list rules: %w", err)
-	}
-
-	desiredState := []rulefmt.RuleGroup{}
-	for _, pr := range crdState {
-		groups, err := convertCRDRuleGroupToRuleGroup(pr.Spec)
-		if err != nil {
-			return fmt.Errorf("failed to convert rule group: %w", err)
-		}
-
-		desiredState = append(desiredState, groups.Groups...)
-	}
-
-	diffs, err := c.diffRuleStates(desiredState, c.currentState)
-	if err != nil {
-		return err
-	}
-
-	return c.applyChanges(ctx, diffs)
-}
-
 func convertSelectorToListOptions(selector LabelSelector) (labels.Selector, error) {
 	matchExpressions := []metav1.LabelSelectorRequirement{}
 
@@ -253,122 +214,6 @@ func convertSelectorToListOptions(selector LabelSelector) (labels.Selector, erro
 		MatchLabels:      selector.MatchLabels,
 		MatchExpressions: matchExpressions,
 	})
-}
-
-type RuleGroupDiffKind string
-
-const (
-	RuleGroupDiffKindAdd    RuleGroupDiffKind = "add"
-	RuleGroupDiffKindRemove RuleGroupDiffKind = "remove"
-	RuleGroupDiffKindUpdate RuleGroupDiffKind = "update"
-)
-
-type RuleGroupDiff struct {
-	Kind    RuleGroupDiffKind
-	Actual  mimirClient.RuleGroup
-	Desired mimirClient.RuleGroup
-}
-
-func (c *Component) diffRuleStates(desired []rulefmt.RuleGroup, actual []mimirClient.RuleGroup) ([]RuleGroupDiff, error) {
-	var diff []RuleGroupDiff
-
-	seenGroups := map[string]bool{}
-
-desiredGroups:
-	for _, desiredRuleGroup := range desired {
-		mimirRuleGroup := mimirClient.RuleGroup{
-			RuleGroup: desiredRuleGroup,
-			// TODO: allow setting the remote write configs?
-			// RWConfigs: ,
-		}
-
-		seenGroups[desiredRuleGroup.Name] = true
-
-		for _, actualRuleGroup := range actual {
-			if desiredRuleGroup.Name == actualRuleGroup.Name {
-				if equalRuleGroups(desiredRuleGroup, actualRuleGroup.RuleGroup) {
-					continue desiredGroups
-				}
-
-				// TODO: check if the rules are the same
-				diff = append(diff, RuleGroupDiff{
-					Kind:    RuleGroupDiffKindUpdate,
-					Actual:  actualRuleGroup,
-					Desired: mimirRuleGroup,
-				})
-				continue desiredGroups
-			}
-		}
-
-		diff = append(diff, RuleGroupDiff{
-			Kind:    RuleGroupDiffKindAdd,
-			Desired: mimirRuleGroup,
-		})
-	}
-
-	for _, actualRuleGroup := range actual {
-		if seenGroups[actualRuleGroup.Name] {
-			continue
-		}
-
-		diff = append(diff, RuleGroupDiff{
-			Kind:   RuleGroupDiffKindRemove,
-			Actual: actualRuleGroup,
-		})
-	}
-
-	return diff, nil
-}
-
-func (c *Component) applyChanges(ctx context.Context, diffs []RuleGroupDiff) error {
-	if len(diffs) == 0 {
-		return nil
-	}
-
-	level.Info(c.log).Log("msg", "applying rule changes", "num_changes", len(diffs))
-
-	for _, diff := range diffs {
-		switch diff.Kind {
-		case RuleGroupDiffKindAdd:
-			level.Info(c.log).Log("msg", "adding rule group", "group", diff.Desired.Name)
-			err := c.mimirClient.CreateRuleGroup(ctx, c.args.MimirRuleNamespace, diff.Desired)
-			if err != nil {
-				return err
-			}
-		case RuleGroupDiffKindRemove:
-			level.Info(c.log).Log("msg", "removing rule group", "group", diff.Actual.Name)
-			err := c.mimirClient.DeleteRuleGroup(ctx, c.args.MimirRuleNamespace, diff.Actual.Name)
-			if err != nil {
-				return err
-			}
-		case RuleGroupDiffKindUpdate:
-			level.Info(c.log).Log("msg", "updating rule group", "group", diff.Desired.Name)
-			err := c.mimirClient.CreateRuleGroup(ctx, c.args.MimirRuleNamespace, diff.Desired)
-			if err != nil {
-				return err
-			}
-		default:
-			level.Error(c.log).Log("msg", "unknown rule group diff kind", "kind", diff.Kind)
-		}
-	}
-
-	c.syncMimir(ctx)
-
-	return nil
-}
-
-func convertCRDRuleGroupToRuleGroup(crd promv1.PrometheusRuleSpec) (*rulefmt.RuleGroups, error) {
-	buf, err := yaml.Marshal(crd)
-	if err != nil {
-		return &rulefmt.RuleGroups{}, err
-	}
-
-	groups, errs := rulefmt.Parse(buf)
-	if len(errs) > 0 {
-		return &rulefmt.RuleGroups{}, multierror.New(errs...).Err()
-	}
-
-	return groups, nil
 }
 
 func (c *Component) startRuleInformer() {
@@ -429,118 +274,8 @@ func (c *Component) startRuleInformer() {
 	factory.WaitForCacheSync(c.informerStopChan)
 }
 
-func (c *Component) eventLoop(ctx context.Context) {
-	for {
-		event, shutdown := c.queue.Get()
-		if shutdown {
-			level.Info(c.log).Log("msg", "shutting down event loop")
-			return
-		}
-
-		evt := event.(Event)
-		err := c.processEvent(ctx, evt)
-
-		if err != nil {
-			// TODO: retry limits?
-			level.Error(c.log).Log("msg", "failed to process event", "err", err)
-			// c.queue.AddRateLimited(event)
-		} else {
-			c.queue.Forget(event)
-			c.queue.Done(event)
-		}
+func setDefaultArguments(args *Arguments) {
+	if args.SyncInterval == 0 {
+		args.SyncInterval = 30 * time.Second
 	}
-}
-
-func (c *Component) getRuleGroupsFromKey(key string) (*rulefmt.RuleGroups, error) {
-	obj, _, err := c.ruleInformer.GetIndexer().GetByKey(key)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to get rule from informer")
-	}
-
-	groups, err := convertCRDRuleGroupToRuleGroup(obj.(*promv1.PrometheusRule).Spec)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to convert CRD rule group to rule group")
-	}
-
-	return groups, nil
-}
-
-func (c *Component) processEvent(ctx context.Context, e Event) error {
-	switch e.Type {
-	case EventTypeAddRule:
-		level.Info(c.log).Log("msg", "processing add rule event", "key", e.NewRules)
-	case EventTypeUpdateRule:
-		level.Info(c.log).Log("msg", "processing update rule event", "key", e.NewRules)
-	case EventTypeDeleteRule:
-		level.Info(c.log).Log("msg", "processing delete rule event", "key", e.OldRules)
-	case EventTypeAddNamespace:
-	case EventTypeDeleteNamespace:
-	case EventTypeUpdateNamespace:
-	case EventTypeSyncMimir:
-		level.Debug(c.log).Log("msg", "syncing current state from ruler")
-		c.syncMimir(ctx)
-	default:
-		return fmt.Errorf("unknown event type: %s", e.Type)
-	}
-
-	return c.reconcileState(ctx)
-}
-
-func (c *Component) syncMimir(ctx context.Context) {
-	rulesByNamespace, err := c.mimirClient.ListRules(ctx, c.args.MimirRuleNamespace)
-	if err != nil {
-		level.Error(c.log).Log("msg", "failed to list rules from mimir", "err", err)
-		return
-	}
-
-	c.currentState = rulesByNamespace[c.args.MimirRuleNamespace]
-}
-
-// This type must be hashable, so it is kept simple. The indexer will maintain a
-// cache of current state, so this is only used for logging.
-type Event struct {
-	Type EventType
-
-	NewRules string
-	OldRules string
-
-	NewNamespace string
-	OldNamespace string
-}
-
-type EventType string
-
-const (
-	EventTypeAddRule    EventType = "add-rule"
-	EventTypeUpdateRule EventType = "update-rule"
-	EventTypeDeleteRule EventType = "delete-rule"
-
-	EventTypeAddNamespace    EventType = "add-namespace"
-	EventTypeUpdateNamespace EventType = "update-namespace"
-	EventTypeDeleteNamespace EventType = "delete-namespace"
-
-	EventTypeSyncMimir EventType = "sync-mimir"
-)
-
-func equalRuleGroups(a, b rulefmt.RuleGroup) bool {
-	aBuf, err := yamlv3.Marshal(a)
-	if err != nil {
-		return false
-	}
-	bBuf, err := yamlv3.Marshal(b)
-	if err != nil {
-		return false
-	}
-
-	if !bytes.Equal(aBuf, bBuf) {
-
-		fmt.Println("----")
-		fmt.Println(string(aBuf))
-		fmt.Println("----")
-		fmt.Println(string(bBuf))
-
-		return false
-	}
-
-	return bytes.Equal(aBuf, bBuf)
 }
